Gravity Upgrade!
How to use:
Drag prefab "Gravity Manager" into the scene, set the volum to G and Threshold. G is the factors used to compute gravity. The larger the G is, the stronger force will apply. Threahold is for preventing the gravity becomes too large which would cause player object "disappear" suddently. Literally, 1000 is the magic number for G whereas Threshold you may want to set a new one.

Script ForceObject should be used on planet (root object). There are Active, Force Type and Force Magnitude. Active is for making the gravity effect for this object, vise versa. Force Type consist of positive and negative, work like actual charged particals: ++/-- = attraction, +- = repel. Force Magnitude is the factor saying the strength of the gavity the object will have.

Script ForceAffectedObject should be used on player object. Actually, it is same to script ForceObject except for the rigidbody which will be set automatically. 1 for Force Magnitude and 1 for mass work normally at current stage. I very recommend you to try different volum for different expriences.

1. ForceManager manages all force-effected objects
2. ForceObject allows a object create "gravity"
3. ForceAffectedObject let a physical object move
4. remove deprecated files